requirements.txt
mcp
yfinance
python-dotenv

